+++ machines/src/gluetex/new.rs
@@ -642,6 +642,7 @@ impl MachineNewTrait for Gluetex {
                 inlet_feeder_tension_arm_monitor: super::TensionArmMonitor::new("InletFeeder"),
                 optris_1_monitor: super::VoltageMonitor::new("Optris 1"),
                 optris_2_monitor: super::VoltageMonitor::new("Optris 2"),
+                bandueberwachung_enabled: true,
                 bandueberwachung_triggered: false,
                 bandueberwachung_not_active_since: None,
                 sleep_timer: super::SleepTimer::new(),